requirements.txt
numpy
pillow
matplotlib
pandas
psutil